Removed redundant uploaders, added attachments, added security for editing and destroy projects and associations

+++ app/controllers/memberships_controller.rb
@@ -1,5 +1,6 @@
 class MembershipsController < ApplicationController
   def create
+    @project = current_account.projects.find(params[:membership][:project_id].to_i)
     @membership = Membership.new(:project => Project.find(params[:membership][:project]))
     @account = Membership.check_if_user(params[:membership])
     unless @account
@@ -15,6 +16,8 @@ class MembershipsController < ApplicationController
         redirect_to :back
       end
     end
+    rescue ActiveRecord::RecordNotFound
+      return redirect_to root_url, :alert => 'You cannot upload attachments this project.'   
   end
 
   def update
@@ -30,12 +33,17 @@ class MembershipsController < ApplicationController
 
   def destroy 
     @membership = Membership.find(params[:id])
-    if @membership.destroy
-      flash[:notice] = "Membership successfully removed"
-      redirect_to :back
-    else
-      flash[:warning] = "Membership not removed"
-      redirect_to :back
+    if current_account.projects.includes(Project.find(@membership.project_id))
+      if @membership.destroy
+        flash[:notice] = "Membership successfully removed"
+        redirect_to :back
+      else
+        flash[:warning] = "Membership not removed"
+        redirect_to :back
+      end
+    else 
+       flash[:warning] = "You cannot remove memberships from this project"
+       redirect_to :root
     end
   end
